Create hangman.cpp
The computer select a word at random from the list of available words that was provided in words.txt. (some 1000s SAT words).
In the hangman.cpp file:
The user is asked to supply one guess (i.e. letter) per round.
The user should receive feedback immediately after each guess about whether their guess appears in the secret word.
After each round, we display the word and the places of the non-guessed characters, as well as the letters that the user has not yet guessed.
We used an array of function pointers to save the drawings.